requirements.txt
aiogram==3.22.0
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.9
aiohttp==3.12.15
pdfplumber==0.11.7
python-docx==1.2.0
python-dotenv==1.1.1
